update: create 9cc.h and divide 9cc.c into several files
