+++ 9cc.h
@@ -1,0 +1,104 @@
+#include <ctype.h>
+#include <stdarg.h>
+#include <stdbool.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+
+/* tokenize.c */
+
+// NOTE: enum は定数を一括で宣言できる型
+typedef enum
+{
+  TK_RESERVED, // 記号
+  TK_NUM,      // 整数トークン
+  TK_EOF       // 入力の終わりを表すトークン
+} TokenKind;
+
+/*
+  NOTE: typedef で struct Token を通常の名前空間で変数宣言できるようにする
+        struct Token { ... };
+        struct Token token; => OK
+        Token token; => Error
+        typedef struct Token token_t;
+        token_t token; => OK
+*/
+typedef struct Token Token;
+
+struct Token
+{
+  TokenKind kind; // トークンの型
+  Token *next;    // 次の入力トークン
+  int val;        // kindがTK_NUMの場合、その数値
+  char *str;      // トークン文字列
+  int len;        // トークンの長さ ※ 数値の長さではない
+};
+
+Token *token;
+void error_at(char *loc, char *fmt, ...);
+void error(char *fmt, ...);
+bool consume(char *op);
+void expect(char *op);
+int expect_number(void);
+bool at_eof(void);
+Token *new_token(TokenKind kind, Token *cur, char *str, int len);
+bool start_with(char *lhs, char *rhs);
+Token *tokenize();
+
+
+/* parse.c */
+
+typedef enum
+{
+  ND_EQ,  // ==
+  ND_NE,  // !=
+  ND_LE,  // <=
+  ND_LT,  // <
+  ND_ADD, // +
+  ND_SUB, // -
+  ND_MUL, // *
+  ND_DIV, // /
+  ND_NUM  // Integer
+} NodeKind;
+
+typedef struct Node Node;
+struct Node
+{
+  NodeKind kind;
+  Node *lhs;
+  Node *rhs;
+  int val;
+};
+
+Node *new_node(NodeKind kind);
+Node *new_binary(NodeKind kind, Node *lhs, Node *rhs);
+Node *new_node_num(int val);
+
+// NODE: 四則演算 + 比較演算は以下で表現される。これをC関数に落とし込む。
+//       expr    = equality
+//       equality = relational ("==" relational | "!=" relational)*
+//       relational = add ("<" add | "<=" add | ">" add | ">=" add)*
+//       add    = mul ("+" mul | "-" mul)*
+//       mul     = unary ("*" unary | "/" unary)*
+//       unary   = ("+" | "-")? primary
+//       primary = num | "(" expr ")"
+Node *expr();
+Node *equality();
+Node *relational();
+Node *add();
+Node *mul();
+Node *unary();
+Node *primary();
+
+
+/* codegen.c */
+
+void gen(Node *node);
+
+
+/* main.c */
+
+char *user_input;
+
+int main(int argc, char **argv);
