+++ parse.c
@@ -1,0 +1,113 @@
+#include "9cc.h"
+
+Node *new_node(NodeKind kind)
+{
+  Node *node = calloc(1, sizeof(Node));
+  node->kind = kind;
+  return node;
+};
+
+Node *new_binary(NodeKind kind, Node *lhs, Node *rhs)
+{
+  Node *node = new_node(kind);
+  node->lhs = lhs;
+  node->rhs = rhs;
+  return node;
+};
+
+Node *new_node_num(int val)
+{
+  Node *node = new_node(ND_NUM);
+  node->val = val;
+  return node;
+}
+
+Node *expr()
+{
+  return equality();
+}
+
+// equality = relational ("==" relational | "!=" relational)*
+Node *equality()
+{
+  Node *node = relational();
+  for (;;)
+  {
+    if (consume("=="))
+      node = new_binary(ND_EQ, node, relational());
+    else if (consume("!="))
+      node = new_binary(ND_NE, node, relational());
+    else
+      return node;
+  }
+}
+
+// relational = add ("<" add | "<=" add | ">" add | ">=" add)*
+Node *relational()
+{
+  Node *node = add();
+  for (;;)
+  {
+    if (consume("<"))
+      node = new_binary(ND_LT, node, add());
+    else if (consume("<="))
+      node = new_binary(ND_LE, node, add());
+    else if (consume(">"))
+      node = new_binary(ND_LT, add(), node);
+    else if (consume(">="))
+      node = new_binary(ND_LE, add(), node);
+    else
+      return node;
+  }
+}
+
+// add = mul ("+" mul | "-" mul)*
+Node *add()
+{
+  Node *node = mul();
+  for (;;)
+  {
+    if (consume("+"))
+      node = new_binary(ND_ADD, node, mul());
+    else if (consume("-"))
+      node = new_binary(ND_SUB, node, mul());
+    else
+      return node;
+  }
+}
+
+Node *mul()
+{
+  Node *node = unary();
+  for (;;)
+  {
+    if (consume("*"))
+      node = new_binary(ND_MUL, node, unary());
+    else if (consume("/"))
+      node = new_binary(ND_DIV, node, unary());
+    else
+      return node;
+  }
+}
+
+Node *unary()
+{
+  if (consume("+"))
+    return primary();
+  else if (consume("-"))
+    return new_binary(ND_SUB, new_node_num(0), primary());
+  else
+    return primary();
+}
+
+Node *primary()
+{
+  if (consume("("))
+  {
+    Node *node = expr();
+    expect(")");
+    return node;
+  }
+
+  return new_node_num(expect_number());
+}
