+++ tokenize.c
@@ -1,0 +1,129 @@
+#include "9cc.h"
+
+// エラーを報告するための関数
+// printfと同じ引数を取る
+// NOTE: ... は引数がいくつあってもよいという意味。
+//       int main() === int main(...) != int main(void)
+void error(char *fmt, ...)
+{
+  va_list ap;
+  va_start(ap, fmt);
+  vfprintf(stderr, fmt, ap);
+  fprintf(stderr, "\n");
+  exit(1);
+}
+
+void error_at(char *loc, char *fmt, ...)
+{
+  va_list ap;
+  va_start(ap, fmt);
+  int pos = loc - user_input;
+  fprintf(stderr, "%s\n", user_input);
+  fprintf(stderr, "%*s", pos, "");
+  fprintf(stderr, "^ ");
+  vfprintf(stderr, fmt, ap);
+  fprintf(stderr, "\n");
+  exit(1);
+}
+
+// 次のトークンが期待している記号のときには、トークンを1つ読み進めて
+// 真を返す。それ以外の場合には偽を返す。
+bool consume(char *op)
+{
+  if (token->kind != TK_RESERVED ||
+      strlen(op) != token->len ||
+      memcmp(token->str, op, token->len))
+    return false;
+  token = token->next;
+  return true;
+}
+
+// 次のトークンが期待している記号のときには、トークンを1つ読み進める。
+// それ以外の場合にはエラーを報告する。
+void expect(char *op)
+{
+  if (token->kind != TK_RESERVED ||
+      strlen(op) != token->len ||
+      memcmp(token->str, op, token->len))
+    error_at(token->str, "'%s'ではありません", op);
+  token = token->next;
+}
+
+// 次のトークンが数値の場合、トークンを1つ読み進めてその数値を返す。
+// それ以外の場合にはエラーを報告する。
+int expect_number(void)
+{
+  if (token->kind != TK_NUM)
+    error_at(token->str, "数値ではありません");
+  int val = token->val;
+  token = token->next;
+  return val;
+}
+
+bool at_eof(void)
+{
+  return token->kind == TK_EOF;
+}
+
+Token *new_token(TokenKind kind, Token *cur, char *str, int len)
+{
+  Token *tok = calloc(1, sizeof(Token));
+  tok->kind = kind;
+  tok->str = str;
+  tok->len = len;
+  cur->next = tok;
+  return tok;
+}
+
+bool start_with(char *lhs, char *rhs)
+{
+  return memcmp(lhs, rhs, strlen(rhs)) == 0;
+}
+
+// 入力文字列user_inputをトークナイズしてそれを返す
+Token *tokenize()
+{
+  char *p = user_input;
+  Token head;
+  head.next = NULL;
+  Token *cur = &head;
+
+  while (*p)
+  {
+    // 空文字はスキップ
+    if (isspace(*p))
+    {
+      p++;
+      continue;
+    }
+
+    // multi-letter punctuator
+    if (start_with(p, "==") || start_with(p, "!=") || start_with(p, ">=") || start_with(p, "<="))
+    {
+      cur = new_token(TK_RESERVED, cur, p, 2);
+      p += 2;
+      continue;
+    }
+
+    // single-letter punctuator
+    if (strchr("+-*/()<>", *p))
+    {
+      cur = new_token(TK_RESERVED, cur, p++, 1);
+      continue;
+    }
+
+    if (isdigit(*p))
+    {
+      cur = new_token(TK_NUM, cur, p, 0);
+      char *q = p;
+      cur->val = strtol(p, &p, 10);
+      cur->len = p - q; // 数値の部分がどこからどこまでは事前にわからないので、次のトークンの位置との差分で算出する
+      continue;
+    }
+
+    error_at(p, "トークナイズできません");
+  }
+
+  new_token(TK_EOF, cur, p, 0);
+  return head.next;
+}
